basic ObservatoryRespositoryViewModel
